open notification on mouse.screen
Notification were occasionally opened on the wrong screen when using the
mouse to hover over a widget.

It is possible to have the mouse on a different screen than the focused
client by actively moving the mouse from screen to screen without
leaving the topbar. Switching screens with the keyboard binding will
also jump the mouse to the new screen. Therefore, using 'mouse.screen'
instead of 'client.focus.screen' will always open the notification on
the correct screen. 'mouse.screen' should always be defined (correct me
if I'm wrong), so no need for 'or 1'.

This was tested for the calendar widget only, the other ones where
replaced with sed and some care but are untested.

Side note:
On the other hand, it might be even better to be able to pass screen as
an argument when creating/showing the widget. This is especially usefull
when showing the widget via keyboard binding.