Optionally disable mapped volume output
The -M option in amixer produces different behaviours in different
hardware. Users may prefer a linear volume scale (without -M) to a
mapped volume level (with -M).

This patch adds a "mapped" argument that allows disabling the -M
parameter.

The default behaviour is to use -M to keep backwards compatibility.