Better implementation of tag_view_nonempty
The Problem:
The issue will current implementation is that it goes to next tag and
stops only if clients on that tag are more than 0.
So now consider a scenario we have clients on *tag 1,2,5*

With the current implementation if we are on tag 2 and call
lain.util.tag_view_nonempty(1) and then call
awful.tag.history.restore() It will take us to tag 4 instead of tag 2.

This PR solves the problem